Fix for incorrect OS Version Detection (#22)
The existing code checks only the middle digit to determine if the machine is running Snow Leopard. Snow Leopard is version 10.6. With Big Sur, there is a version 11.6, which created a false positive.
Updated the OS detection routine to return major.minor and updated the version check to look at "10.6"